Models repositories

+++ src/main/java/com/codewranglers/workflowmanager/models/data/UserRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-
+public interface UserRepository extends CrudRepository<User, Integer> {
 }
